@@ -1,6 +1,6 @@
-// +build tools
-
+// go:build tools
 // Why does this file exists? see https://github.com/golang/go/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module
+
 package tools
 
 import (
